Added client.sendWillQueue() and "readytosend" event.
This is useful for throttling sends. I have rapidly-changing state
on the server, but I only need to send the most recent state to the
client - so if sendWillQueue() is true I wait for the 'readytosend'
event before sending the latest state.

'readytosend' is emitted from _payload() after the connection state
is set up, but before sending the content of the write queue; so
a readytosend callback can call send() and it will be sent immediately.