Use single quotes in htmlfile protocol
In JSON, there are lots of unavoidable double-quotes: For strings, and for the
object keys. Looking at a htmlfile stream you'll find many double quote escapes.

Example:

<script>_("5:::{\"name\":\"update\",\"args\":[{\"Q@G@165120019@O\":{\"GBP8\":9445,\"GBV8\":904,\"s\":1}}]}");</script>

If we used a single quote at the outer level though, we wouldn't need to escape
the double quotes, only the - much less used in JSON - single quotes.

The example above would thus turn into:

<script>_("5:::{"name":"update","args":[{"Q@G@165120019@O":{"GBP8":9445,"GBV8":904,"s":1}}]}");</script>

Nicer to look at, and consumes fewer bytes. IE - the browser htmlfile is targeted
at - is fine with single-quotes, as it's legal JavaScript, and socket.io-spec does
not specify that, so I think it would be OK to change