Base64 decoding support
Added a function that takes the base64 encoding of our event packets and
decodes them into packet json with correct type, namespace, and data args,
where data args are defined as [eventName, {base64: true, data: base64string}].
Then we leave it up to the user to handle something where base64 is true.

This bit of code is only relevant for very old browsers.