Added link to HyperDev version of chat demo
This adds a remix link to a version of the Socket.io Chat demo on HyperDev - some users might prefer this as they get a copy of the code that just runs without any setup, and they can edit it straight away. I can add maintainers to the HyperDev project.