Small change to the useage with Express
I used the express setup with my latest project and stumbled into an error, that was hard to fix, because it was so small and crucial.  
I accidently called `app.listen(PORT)` instead of `server.listen(PORT)`  
This basically leads to working Express functionality and routes, but the `/socket.io` route will not be present and thus the GET request for any client willing to connect gets a 404.  
I hope this small comment may save another one the four hours I spent on tracing my mistake.