chore: bump engine.io-client and downgrade debug
The newer versions need to be transpiled to ES5, so we'll rollback for
now as it breaks IE11.

We'll have to find a way to provide the users who bundle the Engine.IO
(or Socket.IO) client (with webpack for example) with a source code
which does not include debug (or document the usage of
webpack-remove-debug [1]).

[1] https://github.com/johngodley/webpack-remove-debug